# feat: Parallelize Phase2 checksum calculations for massive performance gains