## Summary
- Implemented worker pool pattern for Phase2 checksum calculations
- Added comprehensive benchmark tests to measure performance improvements
- Achieved 24-25x speedup in realistic network scenarios

## Performance Improvements

Based on benchmark tests with simulated S3 API latency:

| Scenario | Serial Time | Parallel Time | Speedup |
|----------|-------------|---------------|---------|
| 10ms latency (same region) | 1.1s | 46ms | **24x** |
| 50ms latency (cross-region) | 5.1s | 204ms | **25x** |

## Implementation Details
- Uses 32 concurrent workers by default
- Maintains exact output ordering (no breaking changes)
- All existing tests pass without modification
- Memory usage slightly increased due to goroutine overhead

## Test plan
- [x] Run existing unit tests
- [x] Create and run benchmark tests
- [ ] Test with large file sets (1000+ files)
- [ ] Monitor memory usage under load

🤖 Generated with [Claude Code](https://claude.ai/code)